Put statistic methods into class Statistic
